Changed int to long for dCache validation check
git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/branches/SriniProbeTest@17023 4e558342-562e-0410-864c-e07659590f8c

+++ osgtest/tests/test_54_gratia.py
@@ -383,7 +383,7 @@ class TestGratia(osgunittest.OSGTestCase):
         print UsedSpace
         
         #Need to assert only after converting string to integers...
-        self.assert_(int(TotalSpace) == (int(FreeSpace) + int(UsedSpace)))
+        self.assert_(long(TotalSpace) == (long(FreeSpace) + long(UsedSpace)))
         os.remove(filename)
         
     #===============================================================================
